+++ packages/server/src/environment.ts
@@ -35,7 +35,7 @@ const DEFAULTS = {
   USE_LOCAL_COMPONENT_LIBS: coreEnv.isDev() || coreEnv.isTest(),
   RECAPTCHA_SESSION_SECONDS: 1800,
   AUTOMATION_MAX_NESTED_LOOPS: 3,
-  AUTOMATION_LOOPS_RESULTS_THRESHOLD: 50,
+  AUTOMATION_MAX_STORED_LOOP_RESULTS: 50,
 }
 
 const QUERY_THREAD_TIMEOUT =
@@ -103,9 +103,9 @@ const environment = {
   AUTOMATION_MAX_NESTED_LOOPS:
     parseIntSafe(process.env.AUTOMATION_MAX_NESTED_LOOPS) ||
     DEFAULTS.AUTOMATION_MAX_NESTED_LOOPS,
-  AUTOMATION_LOOPS_RESULTS_THRESHOLD:
-    parseIntSafe(process.env.AUTOMATION_LOOPS_RESULTS_THRESHOLD) ||
-    DEFAULTS.AUTOMATION_LOOPS_RESULTS_THRESHOLD,
+  AUTOMATION_MAX_STORED_LOOP_RESULTS:
+    parseIntSafe(process.env.AUTOMATION_MAX_STORED_LOOP_RESULTS) ||
+    DEFAULTS.AUTOMATION_MAX_STORED_LOOP_RESULTS,
   PLUGINS_DIR: process.env.PLUGINS_DIR || DEFAULTS.PLUGINS_DIR,
   MAX_IMPORT_SIZE_MB: process.env.MAX_IMPORT_SIZE_MB,
   SESSION_EXPIRY_SECONDS: process.env.SESSION_EXPIRY_SECONDS,
