+++ packages/types/src/documents/app/automation/automation.ts
@@ -5,6 +5,7 @@ import { Table } from "../table"
 import { AutomationStep, AutomationTrigger } from "./schema"
 import { ContextEmitter } from "../../../sdk"
 import { Readable } from "stream"
+import { LoopSummary } from "./StepInputsOutputs"
 
 export enum AutomationIOType {
   OBJECT = "object",
@@ -351,3 +352,10 @@ export const ImageContentTypes = [
   "image/jpg",
   "image/jpeg",
 ]
+
+export interface LoopStorage {
+  results: Record<string, AutomationStepResult[]>
+  summary: LoopSummary
+  nestedSummaries: Record<string, LoopSummary[]>
+  maxStoredResults: number
+}
